+++ docs/changelog.rst
@@ -20,6 +20,10 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Add specialised rendering for a few more box powerline and unicode symbols
   (:pull:`2074` and :pull:`2021`)
 
+- Add a new socket only mode for :opt:`allow_remote_control`. This makes
+  it possible to programs running on the local machine to control kitty
+  but not programs running over ssh.
+
 
 0.14.6 [2019-09-25]
 ---------------------